requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.5
python-socketio==5.10.0
python-dotenv==1.0.0
pydantic==2.5.3
openpyxl==3.1.2
SQLAlchemy==2.0.25
